Get back failed orders action setting

+++ Helper/Order.php
@@ -190,7 +190,7 @@ class Order extends \Magento\Framework\App\Helper\AbstractHelper
             $this->restoreQuote();
 
             // delete order if needed
-            //if ($this->_config->getValue('order_action_failed_payment') == 'delete') {
+            if ($this->_config->getValue('order_action_failed_payment') == 'delete') {
                 if ($this->_registry->registry('isSecureArea')) {
                     $this->_orderRepository->delete($order);
                 } else {
@@ -198,7 +198,7 @@ class Order extends \Magento\Framework\App\Helper\AbstractHelper
                     $this->_orderRepository->delete($order);
                     $this->_registry->unregister('isSecureArea');
                 }
-            //}
+            }
 
             return true;
         }
